Refractored creator.js and fixed some production routing
