+++ src/front-end/components/block.jsx
@@ -235,7 +235,7 @@ let blockTemplate = <template>
 	<link type="text/css" rel="stylesheet" href="./block.css" />
 	<div id="editorIcons" class="paragraphIcons"><img src="../public/resources/plusIcon.png" class="unfocusedIcons"/><img src="../public/resources/sixDotIcon.png" class="unfocusedIcons"/><div id="signifier"></div></div>
 	<div id="checkerContainer" checked=""><div id="taskChecker"></div></div>
-	<div id="textBlock" contenteditable="true" ondrop={()=>false} placeholder='Type "/" to create a block'></div>
+	<div id="textBlock" contentEditable="true" onDrop={()=>false} placeholder='Type "/" to create a block'></div>
 </template>;
 
 export class TextBlock extends HTMLElement {
